Store particle durations in u32

+++ src/game/magic/spells/SpellsLvl03.cpp
@@ -575,7 +575,7 @@ void IceProjectileSpell::Update() {
 				pd->siz = 20.f;
 				float t = std::min(Random::getf(2000.f, 4000.f),
 				              m_duration - ulCurrentTime + Random::getf(0.f, 500.0f));
-				pd->tolive = checked_range_cast<unsigned long>(t);
+				pd->tolive = checked_range_cast<u32>(t);
 				pd->tc = tex_p2;
 				pd->special = FADE_IN_AND_OUT | ROTATING | MODULATE_ROTATION | DISSIPATING;
 				pd->fparam = 0.0000001f;
@@ -591,7 +591,7 @@ void IceProjectileSpell::Update() {
 				pd->siz = 0.5f;
 				float t = std::min(Random::getf(2000.f, 3000.f),
 				              m_duration - ulCurrentTime + Random::getf(0.f, 500.0f));
-				pd->tolive = checked_range_cast<unsigned long>(t);
+				pd->tolive = checked_range_cast<u32>(t);
 				pd->tc = tex_p1;
 				pd->special = FADE_IN_AND_OUT | ROTATING | MODULATE_ROTATION | DISSIPATING;
 				pd->fparam = 0.0000001f;
